+++ benchexec/tools/condtest-extractor.py
@@ -17,7 +17,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import benchexec.tools.condtesttools as condtest
+import benchexec.tools.condtest as condtest
 
 
 class Tool(condtest.Tool):
@@ -29,4 +29,4 @@ class Tool(condtest.Tool):
     _exec_path = "bin/extractor/test_executor"
 
     def name(self):
-        return "condtest-extractor"
+        return "CondTest Extractor"
