+++ benchexec/tools/condtest-instrumenter.py
@@ -1,0 +1,33 @@
+"""
+BenchExec is a framework for reliable benchmarking.
+This file is part of BenchExec.
+
+Copyright (C) 2007-2017  Dirk Beyer
+All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import benchexec.tools.condtest as condtest
+
+
+class Tool(condtest.Tool):
+    """
+    This tool instruments the test criteria in the C file, such that a verifier
+    can be used on it.
+    url: https://gitlab.com/sosy-lab/software/conditional-testing
+    """
+
+    _exec_path = "bin/instrumenter/instrumenter"
+
+    def name(self):
+        return "CondTest Instrumenter"
